ref: moved assessment service into entities folder; fix: compile errors after merge
