+++ src/main/webapp/app/entities/modeling-assessment/index.ts
@@ -1,0 +1,1 @@
+export * from './modeling-assessment.service';
