+++ src/main/webapp/app/modeling-assessment-editor/index.ts
@@ -1,5 +1,4 @@
 export * from './modeling-assessment-editor.component';
 export * from './modeling-assessment-conflict/modeling-assessment-conflict.component';
 export * from './modeling-assessment-editor.route';
-export * from './modeling-assessment.service';
 export * from './modeling-assessment-dashboard/modeling-assessment-dashboard.component';
